[asset-metadata-xyk] move non-bootstrap tests into paralel group
