+++ test/parallel/xyk-pallet.asset.register.test.ts
@@ -5,16 +5,22 @@
  * @group parallel
  */
 import { getApi } from "../../utils/api";
-import { getEnvironmentRequiredVars } from "../../utils/utils";
+import { getEnvironmentRequiredVars, xykErrors } from "../../utils/utils";
 import { User } from "../../utils/User";
 import { Keyring } from "@polkadot/api";
 import { Assets } from "../../utils/Assets";
-import { ExtrinsicResult } from "../../utils/eventListeners";
+import {
+  ExtrinsicResult,
+  findEventData,
+  signSendFinalized,
+} from "../../utils/eventListeners";
 import { getEventResultFromMangataTx } from "../../utils/txHandler";
 import { BN, hexToU8a } from "@polkadot/util";
-import { MangataGenericEvent } from "@mangata-finance/sdk";
+import { BN_ONE, MangataGenericEvent } from "@mangata-finance/sdk";
 import { getNextAssetId } from "../../utils/tx";
 import { setupApi } from "../../utils/setup";
+import { Xyk } from "../../utils/xyk";
+import { MGA_ASSET_ID } from "../../utils/Constants";
 
 const { sudo: sudoUserName } = getEnvironmentRequiredVars();
 jest.setTimeout(1500000);
@@ -221,4 +227,89 @@ test.skip("register asset and then try to register new one with the same locatio
   const userAssetMetaError = await findAssetError(userRegisterNewAsset);
 
   expect(userAssetMetaError.method).toEqual("ConflictingLocation");
+});
+
+test("register asset with xyk disabled and try to create a pool, expect to fail", async () => {
+  const register = Assets.registerAsset(
+    "Disabled Xyk",
+    "Disabled Xyk",
+    10,
+    undefined,
+    undefined,
+    { operationsDisabled: true }
+  );
+  const result = await signSendFinalized(register, sudo);
+  // assetRegistry.RegisteredAsset [8,{"decimals":10,"name":"0x44697361626c65642058796b","symbol":"0x44697361626c65642058796b","existentialDeposit":0,"location":null,"additional":{"xcm":null,"xyk":{"operationsDisabled":true}}}]
+  const assetId = findEventData(
+    result,
+    "assetRegistry.RegisteredAsset"
+  ).assetId;
+
+  await expect(
+    signSendFinalized(
+      Xyk.createPool(assetId, BN_ONE, MGA_ASSET_ID, BN_ONE),
+      testUser1
+    )
+  ).rejects.toEqual(
+    expect.objectContaining({
+      state: ExtrinsicResult.ExtrinsicFailed,
+      data: xykErrors.FunctionNotAvailableForThisToken,
+    })
+  );
+});
+
+test("register asset with xyk undefined and try to create a pool, expect success", async () => {
+  const register = Assets.registerAsset(
+    "None Xyk",
+    "None Xyk",
+    10,
+    undefined,
+    undefined,
+    undefined
+  );
+  const result = await signSendFinalized(register, sudo);
+  // assetRegistry.RegisteredAsset [8,{"decimals":10,"name":"0x44697361626c65642058796b","symbol":"0x44697361626c65642058796b","existentialDeposit":0,"location":null,"additional":{"xcm":null,"xyk":{"operationsDisabled":true}}}]
+  const assetId = findEventData(
+    result,
+    "assetRegistry.RegisteredAsset"
+  ).assetId;
+
+  await expect(
+    signSendFinalized(
+      Xyk.createPool(assetId, BN_ONE, MGA_ASSET_ID, BN_ONE),
+      testUser1
+    )
+  ).resolves.toEqual(
+    expect.objectContaining({
+      state: ExtrinsicResult.ExtrinsicSuccess,
+    })
+  );
+});
+
+test("register asset with xyk enabled and try to create a pool, expect success", async () => {
+  const register = Assets.registerAsset(
+    "None Xyk",
+    "None Xyk",
+    10,
+    undefined,
+    undefined,
+    { operationsDisabled: false }
+  );
+  const result = await signSendFinalized(register, sudo);
+  // assetRegistry.RegisteredAsset [8,{"decimals":10,"name":"0x44697361626c65642058796b","symbol":"0x44697361626c65642058796b","existentialDeposit":0,"location":null,"additional":{"xcm":null,"xyk":{"operationsDisabled":true}}}]
+  const assetId = findEventData(
+    result,
+    "assetRegistry.RegisteredAsset"
+  ).assetId;
+
+  await expect(
+    signSendFinalized(
+      Xyk.createPool(assetId, BN_ONE, MGA_ASSET_ID, BN_ONE),
+      testUser1
+    )
+  ).resolves.toEqual(
+    expect.objectContaining({
+      state: ExtrinsicResult.ExtrinsicSuccess,
+    })
+  );
 });
